Adds a timer0 module.
This took a three days to get right. At least enough for the test case...

One byte per timer, plus one global is nice. Additionally, FSR is used
to keep track of the next expiring timer while in the Timer0
ISR. Since the point of this module is to handle timers in the idle
loop, I think that's acceptable.